Extract /languages/... to separate module
